# dispatcher_microservice/handler/request_body_structs.go
//     This is Factory Games Organizer api. Api is responsible for creating, updating and authenicating api users, CRUD operations on database associated with the api and provides production calculator service.
//     Copyright (C) 2025  Marek Banaś

//     This program is free software: you can redistribute it and/or modify
//     it under the terms of the GNU General Public License as published by
//     the Free Software Foundation, either version 3 of the License, or
//     (at your option) any later version.

//     This program is distributed in the hope that it will be useful,
//     but WITHOUT ANY WARRANTY; without even the implied warranty of
//     MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
//     GNU General Public License for more details.

//     You should have received a copy of the GNU General Public License
//     along with this program.  If not, see https://www.gnu.org/licenses/.

package handler

type JSONDataUsers struct {
	UserLogin    string
	UserPassword string
}

type JSONDataCrud struct {
	MachinesList        []MachineInfo
	ResourcesList       []ResourceInfo
	RecipesList         []RecipeInfo
	RecipesInputsList   []RecipeInputOutputInfo
	RecipesOutputsList  []RecipeInputOutputInfo
	MachinesRecipesList []MachinesRecipesInfo
}

type MachineInfo struct {
	Id                 uint
	Name               string
	UsersId            uint
	InputsSolid        uint
	InputsLiquid       uint
	OutputsSolid       uint
	OutputsLiquid      uint
	Speed              float32
	PowerConsumptionKw uint
	DefaultChoice      uint8
}

type MachinesRecipesInfo struct {
	Id         uint
	UsersId    uint
	RecipesId  uint
	MachinesId uint
}

type RecipeInfo struct {
	Id              uint
	Name            string
	UsersId         uint
	ProductionTimeS uint
	DefaultChoice   uint8
}

type RecipeInputOutputInfo struct {
	Id          uint
	UsersId     uint
	RecipesId   uint
	ResourcesId uint
	Amount      uint
}

type ResourceInfo struct {
	Id           uint
	Name         string
	UsersId      uint
	Liquid       uint8
	ResourceUnit string
}

type DeleteInputCrud struct {
	MachinesIds        []int
	ResourcesIds       []int
	RecipesIds         []int
	RecipesInputsIds   []int
	RecipesOutputsIds  []int
	MachinesRecipesIds []int
}